Basic necessary information on Monopoly board
Justin Tomlinson
August 19, 2019

I have implemented some classes for the different tiles in Monopoly and then
the Monopoly class contains a board object which is populated with all of the
tile information. This will be the foundation of the game / simulation which
will be the board just with rules / AI implemented.