Started some of the actual functionality of the game
Justin Tomlinson
August 19, 2019

I've began to work on the gameplay of monopoly which can be used
by the AI that I will implement later. I am also trying to keep it
versatile so that someone could use it as a basis for an actual
human player(s) as well if they wanted to.